Fixed some bugs with the backend of Facebook Connect (refactored at LcAuth), added some utilities required (LcExtensions), and implemented the #538 by integrating account creation through Facebook Connect in the WelcomePopup rather than require a redirect to the Account/Facebook page, filling form fields with Facebook account data (and additional info for the profile that comes from FB), hidding most fields but still user needs to confirm or fill the remaining.

+++ web/App_Code/LcAuth.cs
@@ -115,7 +115,7 @@ public static class LcAuth
             LcMessaging.SendWelcomeCustomer(user.UserID, user.Email, confirmationUrl, user.ConfirmationToken);
     }
 
-    public static void ConnectWithFacebookAccount(int userID, int facebookID)
+    public static void ConnectWithFacebookAccount(int userID, long facebookID)
     {
         using (var db = Database.Open("sqlloco")){
